added buttons for gravity on every selection

+++ src/editor_env_ceilling_buttons.c
@@ -6,7 +6,7 @@
 /*   By: sipatry <sipatry@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/16 11:24:25 by sipatry           #+#    #+#             */
-/*   Updated: 2020/02/03 17:25:21 by sipatry          ###   ########.fr       */
+/*   Updated: 2020/02/03 17:48:39 by sipatry          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ void	init_ceilling_sector_env(t_env *env)
 	env->editor.hud.s_ceilling.t_intensity.type = INT;
 	env->editor.hud.s_ceilling.t_gravity.env = env;
 	env->editor.hud.s_ceilling.t_gravity.pos = new_point(0, 0);
-	env->editor.hud.s_ceilling.t_gravity.type = INT;
+	env->editor.hud.s_ceilling.t_gravity.type = DOUBLE;
 }
 
 void	init_ceilling_general_env(t_env *env)
