+++ src/editor_env_object_buttons.c
@@ -6,7 +6,7 @@
 /*   By: sipatry <sipatry@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/22 15:55:03 by sipatry           #+#    #+#             */
-/*   Updated: 2020/01/24 15:21:51 by sipatry          ###   ########.fr       */
+/*   Updated: 2020/02/03 18:02:43 by sipatry          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ void	init_object_sector_env(t_env *env)
 	env->editor.hud.s_object.t_intensity.type = INT;
 	env->editor.hud.s_object.t_gravity.env = env;
 	env->editor.hud.s_object.t_gravity.pos = new_point(0, 0);
-	env->editor.hud.s_object.t_gravity.type = INT;
+	env->editor.hud.s_object.t_gravity.type = DOUBLE;
 }
 
 void	init_object_general_env(t_env *env)
