+++ src/editor_env_wall_buttons.c
@@ -6,7 +6,7 @@
 /*   By: sipatry <sipatry@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/15 17:46:44 by sipatry           #+#    #+#             */
-/*   Updated: 2020/02/03 17:17:26 by sipatry          ###   ########.fr       */
+/*   Updated: 2020/02/03 18:22:12 by sipatry          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ void	init_wall_sector_env(t_env *env)
 	env->editor.hud.s_wall.t_intensity.type = INT;
 	env->editor.hud.s_wall.t_gravity.env = env;
 	env->editor.hud.s_wall.t_gravity.pos = new_point(0, 0);
-	env->editor.hud.s_wall.t_gravity.type = INT;
+	env->editor.hud.s_wall.t_gravity.type = DOUBLE;
 }
 
 void	init_wall_general_env(t_env *env)
